Create Aboutpage.php
It includes details about the assignment. It provide a brief description ( assignment #1 for COMP3512 at Mount Royal University), styled nicely, about what this site is about, what technologies using, the author name, and the URL for the github repo. Notice also there are buttons or links that will appear on every PHP page.